extract rootReducer into separate module for clarity
