+++ src/modules/main/components/ImportantContainer.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import Carousel from '../../../lib/components/Carousel';
 import { PostCard } from '../../../lib/components/PostCard';
-import { RootState } from '../../../store/store';
+import { RootState } from '../../../store/rootReducer';
 import { loadImportant } from '../store/mainSlice';
 
 const selectImportant = (state: RootState) => state.main.important;
