+++ src/store/store.ts
@@ -1,12 +1,14 @@
 import { configureStore } from '@reduxjs/toolkit';
-import mainReducer from '../modules/main/store/mainSlice';
+import rootReducer from './rootReducer';
 
 export const store = configureStore({
-  /* to create a root reducer combineReducers() is called behind the scenes */
-  reducer: {
-    main: mainReducer,
-    // direction: directionReducer,
-  },
+  reducer: rootReducer,
+  /* we can also pass and object with reducers, the to create a root reducer
+   * combineReducers() will be called behind the scenes */
+  // reducer: {
+  //   main: mainReducer,
+  //   direction: directionReducer,
+  // },
   /* these are passed to applyMiddleware()
    * if not provided, getDefaultMiddleware is called, the defaults are
    * [thunk, immutableStateInvariant, serializableStateInvariant] */
@@ -14,7 +16,5 @@ export const store = configureStore({
   /* Redux DevTools Extension is enabled by default */
   // devTools: true,
 });
-
-export type RootState = ReturnType<typeof store.getState>;
 
 export type AppDispatch = typeof store.dispatch;
